+++ tests/files/src/0779_const_suggestion.php
@@ -1,0 +1,5 @@
+<?php
+echo PHP_VERSION_I;
+echo php_version_id;
+echo fals;
+echo NUL;